pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "DUOLabs333", email = "dvdugo333@gmail.com"},
]
dependencies = ["pyspellchecker==0.4.0", "flask==1.0.3", "cherrypy==18.1.2", "requests==2.22.0"]
requires-python = ">=3.9"
dynamic = ["classifiers"]
license = {text = "MIT"}

[project.urls]
homepage = ""

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"


[tool]
[tool.pdm]
[[tool.pdm.source]]
name = "pypi"
url = "https://pypi.org/simple"
verify_ssl = true
